+++ src/Humanizer.Tests/Localisation/es/NumberToWordsTests.cs
@@ -43,9 +43,20 @@ namespace Humanizer.Tests.Localisation.es
         [InlineData(1999, "mil novecientos noventa y nueve")]
         [InlineData(2014, "dos mil catorce")]
         [InlineData(2048, "dos mil cuarenta y ocho")]
-        public void ToWordsSpanish(int number, string expected)
+        public void ToWords(int number, string expected)
         {
             Assert.Equal(expected, number.ToWords());
+        }
+
+        [Theory]
+        [InlineData(1, "primero", null)]
+        [InlineData(2, "segundo", GrammaticalGender.Masculine)]
+        [InlineData(2, "segunda", GrammaticalGender.Feminine)]
+        [InlineData(2, "segundo", GrammaticalGender.Neuter)]
+        [InlineData(11, "once", null)]
+        public void ToOrdinalWords(int number, string words, GrammaticalGender gender)
+        {
+            Assert.Equal(words, number.ToOrdinalWords(gender));
         }
     }
 }